Refactor: extract code from _reader.py to _cell_iteration.py, _constants.py, dump.py and util.py
