+++ scripts/parse-and-dump.py
@@ -7,9 +7,9 @@ from pathlib import Path
 
 from mireport.cli import configure_rich_output
 from mireport.cli import console_print as print
-from mireport.xlsx_template_reader._reader import (
+from mireport.xlsx_template_reader.dump import getNamedRanges
+from mireport.xlsx_template_reader.util import (
     checkExcelFilePath,
-    getNamedRanges,
     loadExcelFromPathOrFileLike,
 )
