+++ src/mireport/xlsx_template_reader/processor.py
@@ -28,10 +28,12 @@ from mireport.taxonomy import (
 )
 from mireport.version import OUR_VERSION_HOLDER, VersionHolder
 from mireport.xlsx_template_reader._bindings import WorkbookBindings
-from mireport.xlsx_template_reader._fact_creator import FactCreator
-from mireport.xlsx_template_reader._reader import (
+from mireport.xlsx_template_reader._constants import (
     EXCEL_VALUES_TO_BE_TREATED_AS_NONE_VALUE,
-    WorkbookReader,
+)
+from mireport.xlsx_template_reader._fact_creator import FactCreator
+from mireport.xlsx_template_reader._reader import WorkbookReader
+from mireport.xlsx_template_reader.util import (
     excelDefinedNameRef,
     loadExcelFromPathOrFileLike,
 )
